🚀moved explore bar and some other files in wrong atomic order
