+++ Frontend/src/app/app.routes.ts
@@ -2,7 +2,7 @@ import { RouterModule, Routes } from "@angular/router";
 import { LandingPageComponent } from "./pages/landing-page/landing-page.component";
 import { LoginComponent } from "./pages/login/login.component";
 import { RegisterComponent } from "./pages/register/register.component";
-import { HomeComponent } from "./pages/home/home.component";
+import { HomeComponent } from "./components/templates/desktop/home/home.component";
 import { ProfileComponent } from "./pages/profile/profile.component";
 import { AuthCallbackComponent } from "./authcallback/authcallback.component";
 import { UserLibraryComponent } from "./pages/user-library/user-library.component";
