+++ Frontend/src/app/components/organisms/explore-bar/explore-bar.component.spec.ts
