+++ Frontend/src/app/components/templates/desktop/home/home.component.html
@@ -1,0 +1,3 @@
+
+<app-explore-bar class="col-start-2 col-span-5 row-start-2 row-span-1"></app-explore-bar>
+<app-moods></app-moods>
