+++ Frontend/src/app/components/templates/desktop/settings/display/display.component.spec.ts
