--- Frontend/src/app/pages/home/home.component.html
@@ -1,1 +1,0 @@
-<app-moods></app-moods>
